requirements.txt
Django
python-decouple
pandas
openpyxl
google-generativeai
Pillow
psycopg2-binary
django-import-export
